XWIKI-10140: Improve the performance of the Solr Index Job * Fix for Java 6

+++ xwiki-platform-core/xwiki-platform-search/xwiki-platform-search-solr/xwiki-platform-search-solr-api/src/main/java/org/xwiki/search/solr/internal/job/IndexerJob.java
@@ -336,8 +336,8 @@ public class IndexerJob extends AbstractJob<IndexerRequest, DefaultJobStatus<Ind
         Query selectQuery = queries[1];
 
         // We need the count query only to be able to display the job progress.
-        long documentCount = (long) countQuery.setWiki(wiki).execute().get(0);
-        notifyPushLevelProgress((int) documentCount);
+        Long documentCount = (Long) countQuery.setWiki(wiki).execute().get(0);
+        notifyPushLevelProgress(documentCount.intValue());
 
         SolrInstance solrInstance = this.solrInstanceProvider.get();
         selectQuery.setWiki(wiki).setLimit(LIMIT).setOffset(0);
